backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dermatrace-backend"
version = "0.1.0"
description = "DermaTrace FastAPI backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy[asyncio]",
    "asyncpg",
    "alembic",
    "pydantic[email]",
    "pydantic-settings",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "python-multipart",
    "pandas",
    "scikit-learn",
    "stripe",
    "httpx",
    "slowapi",
    "supabase",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "hypothesis",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["app/tests"]
